rename StepPage to FormPage and move it to a general component directory
